refactor: separation of forum category components
