+++ src/components/categoryfilter/PrimaryCategories.tsx
@@ -1,11 +1,5 @@
-import Down from '@/assets/images/common/Down';
-import React, { useEffect, useRef, useState } from 'react';
 import CategoryButton from './CategoryButton';
-
-type primaryForumOption = {
-  type: '전체' | '일상' | '커리어' | '자기개발' | '토론' | '코드리뷰';
-  label: string;
-};
+import ForumToggle from './ForumToggle';
 
 type PrimaryCategoriesProps = {
   primaryCategory: 'all' | 'qna' | 'forum' | 'archive';
@@ -20,99 +14,20 @@ const PrimaryCategories = ({
   onCategoryChange,
   onForumCategoryChange
 }: PrimaryCategoriesProps) => {
-  const [showForumMenu, setShowForumMenu] = useState(false);
-
-  const forumMenuRef = useRef<HTMLDivElement | null>(null);
-
-  useEffect(() => {
-    const handleForumMenuClickOutside = (event: MouseEvent) => {
-      if (forumMenuRef.current && !forumMenuRef.current.contains(event.target as Node)) {
-        setShowForumMenu(false);
-      }
-    };
-
-    document.addEventListener('mousedown', handleForumMenuClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleForumMenuClickOutside);
-    };
-  }, []);
-
-  const toggleForumMenu = (e: React.MouseEvent<HTMLButtonElement>) => {
-    setShowForumMenu(!showForumMenu);
-    const value = (e.target as HTMLButtonElement).value;
-    if (value === '전체') {
-      onCategoryChange('forum');
-      onForumCategoryChange(null);
-    } else if (value) {
-      onCategoryChange('forum');
-      onForumCategoryChange(value);
-    } else {
-      onCategoryChange('forum');
-    }
-  };
-
   const handleCategoryClick = (category: 'all' | 'qna' | 'forum' | 'archive') => {
     onCategoryChange(category);
     onForumCategoryChange(null);
   };
 
-  const handleForumCategoryClick = (type: '전체' | '일상' | '커리어' | '자기개발' | '토론' | '코드리뷰') => {
-    onForumCategoryChange(type);
-    setShowForumMenu(false);
-  };
-
-  const forumCategories: primaryForumOption[] = [
-    { type: '전체', label: '포럼' },
-    { type: '일상', label: '일상' },
-    { type: '커리어', label: '커리어' },
-    { type: '자기개발', label: '자기개발' },
-    { type: '토론', label: '토론' },
-    { type: '코드리뷰', label: '코드리뷰' }
-  ];
-
-  const currentForumCategoryLabel =
-    forumCategories.find((category) => category.type === primaryForumCategory)?.label || '포럼';
-
   return (
     <div className="flex">
       <CategoryButton category="all" currentCategory={primaryCategory} onClick={() => handleCategoryClick('all')} />
-      <div>
-        <button
-          onClick={toggleForumMenu}
-          className={`w-[75px] md:w-[118px] h-[40px] p-[8px_16px] mr-6 flex items-center justify-between text-subtitle3 md:text-subtitle1 font-medium ${
-            primaryCategory === 'forum' ? 'filter-active' : 'filter-inactive'
-          }`}
-        >
-          <p className="w-[64px]">{currentForumCategoryLabel}</p>
-          <Down />
-        </button>
-
-        {showForumMenu && (
-          <div
-            ref={forumMenuRef}
-            className="absolute z-[1000] top-[23px] md:top-[27px] w-[75px] md:w-[118px] border border-neutral-100 rounded-lg bg-white  hover:border hover:border-main-400"
-          >
-            <li
-              onClick={() => handleForumCategoryClick('전체')}
-              className={`flex items-center justify-center p-[8px_16px] h-[40px] text-subtitle3 md:text-subtitle1 font-medium  ${
-                primaryForumCategory === '전체' ? ' text-main-400 ' : ' text-neutral-700'
-              } cursor-pointer`}
-            >
-              <span className="md:mr-2">포럼</span>
-              <Down strokeWidth={1.5} stroke={`${primaryForumCategory === '전체' ? '#423edf' : '#0F0F0F'}`} />
-            </li>
-            {forumCategories.map((option) => (
-              <li
-                key={option.type}
-                onClick={() => handleForumCategoryClick(option.type)}
-                className={`filters-item ${option.type ? 'hover:text-main-400 hover:bg-main-50' : 'text-neutral-700'}`}
-              >
-                {option.type}
-              </li>
-            ))}
-          </div>
-        )}
-      </div>
+      <ForumToggle
+        primaryCategory={primaryCategory}
+        primaryForumCategory={primaryForumCategory}
+        onCategoryChange={onCategoryChange}
+        onForumCategoryChange={onForumCategoryChange}
+      />
       <CategoryButton category="qna" currentCategory={primaryCategory} onClick={() => handleCategoryClick('qna')} />
       <CategoryButton
         category="archive"
